make xdg-runtime-dir thing more robust

+++ system/profile.d/xdg-runtime-dir.sh
@@ -1,9 +1,12 @@
 # Create $XDG_RUNTIME_DIR/shell
 if [ "$XDG_RUNTIME_DIR" ]; then
-    # create directory for the shell
-    mkdir -m 700 "$XDG_RUNTIME_DIR/shell"
+    # Create directory for the shell
+    if [ ! -e "$XDG_RUNTIME_DIR/shell" ]; then
+        mkdir -m 700 "$XDG_RUNTIME_DIR/shell"
+    fi
 
-    # link the XDG_RUNTIME_DIR to $HOME
-    ln -sfT "$XDG_RUNTIME_DIR" "$HOME/.runtime"
+    # Link the XDG_RUNTIME_DIR to $HOME
+    if [ ! -e "$HOME/.runtime" ]; then
+        ln -sfT "$XDG_RUNTIME_DIR" "$HOME/.runtime"
+    fi
 fi
-
